requirements.txt
detectron2==0.6
easydict==1.10
imageio==2.25.0
ipdb==0.13.11
matplotlib==3.6.3
mediapipe==0.9.0.1
mediapipe_silicon==0.8.11
motmetrics==1.4.0
moviepy==1.0.3
numpy==1.23.2
opencv_contrib_python==4.7.0.68
pandas==1.5.3
PyYAML==6.0
scipy==1.10.0
setuptools==65.6.3
tensorflow==2.11.0
tensorflow_macos==2.11.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
ultralytics==8.0.17
git+https://github.com/facebookresearch/detectron2@main#subdirectory=projects/DensePose